refactor: State系のクラスを別ファイルに切り出し
